chore: rename registry package to auth
