+++ dockerconfig/auth/registry.go
@@ -1,4 +1,4 @@
-package registry
+package auth
 
 import (
 	"fmt"
